backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
celery==5.5.3
redis==6.4.0
cryptography==45.0.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
whitenoise==6.9.0
watchdog==6.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.12.0
sqlparse==0.5.5
gunicorn==23.0.0
jsonschema==4.19.1
openai==1.109.1
anthropic==0.66.0
google-generativeai==0.8.5
